Static entrypoints for API
- Separation of static methods and instances to different classes.
- README Migration hints added
- Target version 4.2.0
