+++ README.md
@@ -673,6 +673,11 @@ This workaround was tested and should work for most cases. Please file an issue 
 
 # Migration guide
 
+## Sidenote for updating from 4.1.xx to 4.2.xx
+
+The static entrypoints for the ReMap API changed. Like in earlier versions, you can access the mapping and assertion API with `Mapping.from(...)` and `AssertMapping.of(...)` but if your code held references to instances of those types, you have to change the references to the type `MappingConfiguration`
+respectively `AssertConfiguration`.
+
 ## Sidenote for updating to 4.1.14
 
 The integration of the library propertypath (https://github.com/remondis-it/propertypath) was updated to version 0.1.0.
