+++ src/main/java/com/remondis/remap/MapTransformation.java
@@ -16,7 +16,7 @@ public class MapTransformation extends ReassignTransformation {
 
   private static final String MAP_MSG = "Map %s\n   to %s";
 
-  MapTransformation(Mapping<?, ?> mapping, PropertyDescriptor sourceProperty, PropertyDescriptor destinationProperty) {
+  MapTransformation(MappingConfiguration<?, ?> mapping, PropertyDescriptor sourceProperty, PropertyDescriptor destinationProperty) {
     super(mapping, sourceProperty, destinationProperty);
     denyReassign(sourceProperty, destinationProperty);
   }
