+++ src/main/java/com/remondis/remap/PropertyPathCollectionTransformation.java
@@ -29,13 +29,13 @@ public class PropertyPathCollectionTransformation<RS, X, RD> extends Transformat
   private Get<RS, RD, ?> propertyPath;
   private boolean hasTransformation;
 
-  PropertyPathCollectionTransformation(Mapping<?, ?> mapping, PropertyDescriptor sourceProperty,
+  PropertyPathCollectionTransformation(MappingConfiguration<?, ?> mapping, PropertyDescriptor sourceProperty,
       PropertyDescriptor destinationProperty, PropertyPath<RD, RS, ?> propertyPath) {
     super(mapping, sourceProperty, destinationProperty);
     this.propertyPath = createGetter(sourceProperty, propertyPath);
   }
 
-  PropertyPathCollectionTransformation(Mapping<?, ?> mapping, PropertyDescriptor sourceProperty,
+  PropertyPathCollectionTransformation(MappingConfiguration<?, ?> mapping, PropertyDescriptor sourceProperty,
       PropertyDescriptor destinationProperty, PropertyPath<X, RS, ?> propertyPath, Function<X, RD> transformation) {
     super(mapping, sourceProperty, destinationProperty);
     this.propertyPath = createGetterAndApply(sourceProperty, propertyPath, transformation);
