+++ src/main/java/com/remondis/remap/ReplaceCollectionBuilder.java
@@ -23,10 +23,10 @@ public class ReplaceCollectionBuilder<S, D, RD, RS> {
 
   private TypedPropertyDescriptor<Collection<RS>> sourceProperty;
   private TypedPropertyDescriptor<Collection<RD>> destProperty;
-  private Mapping<S, D> mapping;
+  private MappingConfiguration<S, D> mapping;
 
   ReplaceCollectionBuilder(TypedPropertyDescriptor<Collection<RS>> sourceProperty,
-      TypedPropertyDescriptor<Collection<RD>> destProperty, Mapping<S, D> mapping) {
+      TypedPropertyDescriptor<Collection<RD>> destProperty, MappingConfiguration<S, D> mapping) {
     super();
     this.sourceProperty = sourceProperty;
     this.destProperty = destProperty;
@@ -57,9 +57,9 @@ public class ReplaceCollectionBuilder<S, D, RD, RS> {
    * @param propertyPath A lambda function performing get calls on the specified object to declare the actual property
    *        path.
    *        <b>This is not a function operating on real object. So do not manipulate or calculate here!</b>
-   * @return Returns the {@link Mapping} for further mapping configuration.
+   * @return Returns the {@link MappingConfiguration} for further mapping configuration.
    */
-  public <E extends Exception> Mapping<S, D> withPropertyPath(PropertyPath<RD, RS, E> propertyPath) {
+  public <E extends Exception> MappingConfiguration<S, D> withPropertyPath(PropertyPath<RD, RS, E> propertyPath) {
     denyNull("propertyPath", propertyPath);
     PropertyPathCollectionTransformation<RS, RD, RD> replace = new PropertyPathCollectionTransformation<RS, RD, RD>(
         mapping, sourceProperty.property, destProperty.property, propertyPath);
@@ -92,9 +92,9 @@ public class ReplaceCollectionBuilder<S, D, RD, RS> {
    * #withSkipWhenNull(Function)} to skip on <code>null</code> items.</b>
    *
    * @param transformation The transform function.
-   * @return Returns the {@link Mapping} for further mapping configuration.
+   * @return Returns the {@link MappingConfiguration} for further mapping configuration.
    */
-  public Mapping<S, D> with(Function<RS, RD> transformation) {
+  public MappingConfiguration<S, D> with(Function<RS, RD> transformation) {
     denyNull("tranformation", transformation);
     ReplaceCollectionTransformation<RS, RD> replace = new ReplaceCollectionTransformation<>(mapping,
         sourceProperty.property, destProperty.property, transformation, false);
@@ -108,9 +108,9 @@ public class ReplaceCollectionBuilder<S, D, RD, RS> {
    * skips the execution of the transform function if the source value is <code>null</code>.</b>
    *
    * @param transformation The transform function.
-   * @return Returns the {@link Mapping} for further mapping configuration.
+   * @return Returns the {@link MappingConfiguration} for further mapping configuration.
    */
-  public Mapping<S, D> withSkipWhenNull(Function<RS, RD> transformation) {
+  public MappingConfiguration<S, D> withSkipWhenNull(Function<RS, RD> transformation) {
     denyNull("tranformation", transformation);
     ReplaceCollectionTransformation<RS, RD> replace = new ReplaceCollectionTransformation<>(mapping,
         sourceProperty.property, destProperty.property, transformation, true);
