+++ src/main/java/com/remondis/remap/ReplaceCollectionTransformation.java
@@ -27,7 +27,7 @@ class ReplaceCollectionTransformation<RS, RD> extends SkipWhenNullTransformation
   private Function<RS, RD> transformation;
   private boolean skipWhenNull;
 
-  ReplaceCollectionTransformation(Mapping<?, ?> mapping, PropertyDescriptor sourceProperty,
+  ReplaceCollectionTransformation(MappingConfiguration<?, ?> mapping, PropertyDescriptor sourceProperty,
       PropertyDescriptor destProperty, Function<RS, RD> transformation, boolean skipWhenNull) {
     super(mapping, sourceProperty, destProperty);
     this.transformation = transformation;
